+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/Constants.java
@@ -46,6 +46,7 @@ public final class Constants {
 	// ----- Fixed parameters
 	public static final String ROOT_GROUP_NAME = "RACINE";
 	public static final String ROOT_IDENTIFIER_NAME = "IdUE";
+	public static final String REPORTING_DATA_GROUP_NAME = "REPORTINGDATA";
 
 	// ----- Explicit Variables Names
 	public static final String MODE_VARIABLE_NAME = "MODE_KRAFTWERK";
@@ -63,30 +64,51 @@ public final class Constants {
 	public static final String NUMBER_ORCHESTRATORS_NAME = "NB_ORCHESTRATORS";
 	public static final String NUMBER_SESSIONS_NAME = "NB_SESSIONS";
 	public static final String PARADATA_VARIABLES_PREFIX = "CHANGES_";
+	public static final String PARADATA_TIMESTAMP_SUFFIX = "_LONG";
+	public static final String SURVEY_VALIDATION_DATE_NAME = "DATE_COLLECTE";
+
+	// ----- Paradata event name
+	public static final String PARADATA_SURVEY_VALIDATION_EVENT_NAME = "agree-sending-modal-button-orchestrator-collect";
+
 
 
 	// ----- Reporting Variables Names
+	public static final String REPORTING_DATA_PREFIX_NAME = "Report_";
 	public static final String STATE_SUFFIX_NAME = "STATE";
 	public static final String LAST_STATE_NAME = "LAST_" + Constants.STATE_SUFFIX_NAME;
 	  
-	  public static final String INTERVIEWER_ID_NAME = "IDENQ";
-	  public static final String ORGANIZATION_UNIT_ID_NAME = "ORGANIZATION_UNIT_ID";
-	  public static final String ADRESS_RGES_NAME = "RGES";
-	  public static final String ADRESS_NUMFA_NAME = "NUMFA";
-	  public static final String ADRESS_SSECH_NAME = "SSECH";
-	  public static final String ADRESS_LE_NAME = "LE";
-	  public static final String ADRESS_EC_NAME = "EC";
-	  public static final String ADRESS_BS_NAME = "BS";
-	  public static final String ADRESS_NOI_NAME = "NOI";
-	  public static final String SURVEY_DATE_DAY_NAME = "JOURENQ";
-	  public static final String SURVEY_DATE_MONTH_NAME = "MOISENQ";
-	  public static final String SURVEY_DATE_YEAR_NAME = "ANNEENQ";
-	  public static final String OUTCOME_NAME = "OUTCOME";
-	  public static final String NUMBER_ATTEMPTS_NAME = "NUMBER_CONTACT_ATTEMPTS";
-	  public static final String OUTCOME_ATTEMPT_SUFFIX_NAME = "ATTEMPT";
-	  public static final String ADRESS_ID_STAT_INSEE = "IDSTATINSEE";
+	public static final String INTERVIEWER_ID_NAME = "IDENQ";
+	public static final String ORGANIZATION_UNIT_ID_NAME = "ORGANIZATION_UNIT_ID";
+	public static final String ADRESS_RGES_NAME = "RGES";
+	public static final String ADRESS_NUMFA_NAME = "NUMFA";
+	public static final String ADRESS_SSECH_NAME = "SSECH";
+	public static final String ADRESS_LE_NAME = "LE";
+	public static final String ADRESS_EC_NAME = "EC";
+	public static final String ADRESS_BS_NAME = "BS";
+	public static final String ADRESS_NOI_NAME = "NOI";
+	public static final String SURVEY_DATE_DAY_NAME = "JOURENQ";
+	public static final String SURVEY_DATE_MONTH_NAME = "MOISENQ";
+	public static final String SURVEY_DATE_YEAR_NAME = "ANNEENQ";
+	public static final String OUTCOME_NAME = "OUTCOME";
+	public static final String NUMBER_ATTEMPTS_NAME = "NUMBER_CONTACT_ATTEMPTS";
+	public static final String OUTCOME_ATTEMPT_SUFFIX_NAME = "ATTEMPT";
+	public static final String ADRESS_ID_STAT_INSEE = "IDSTATINSEE";
 	public static final String LAST_ATTEMPT_DATE = "LAST_ATTEMPT_DATE";
 	public static final String OUTCOME_DATE = "OUTCOME_DATE";
+	public static final String IDENTIFICATION_NAME = "identification";
+	public static final String ACCESS_NAME = "access";
+	public static final String SITUATION_NAME = "situation";
+	public static final String CATEGORY_NAME = "category";
+	public static final String OCCUPANT_NAME = "occupant";
+	
+	// ------ Pairwise variables
+	
+	public static final int MAX_LINKS_ALLOWED = 21;
+	public static final String BOUCLE_PRENOMS = "BOUCLE_PRENOMS";
+	public static final String LIEN = "LIEN_";
+	public static final String LIENS = "LIENS";
+	public static final String SAME_AXIS_VALUE = "0";
+	public static final String NO_PAIRWISE_VALUE = "99";
 	
 	// ---------- Functions
 	// ---------- Get a file