+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/extradata/reportingdata/XMLReportingDataParser.java
@@ -2,6 +2,7 @@ package fr.insee.kraftwerk.core.extradata.reportingdata;
 
 import java.nio.file.Path;
 
+import fr.insee.kraftwerk.core.Constants;
 import fr.insee.kraftwerk.core.exceptions.NullException;
 import fr.insee.kraftwerk.core.rawdata.SurveyRawData;
 import fr.insee.kraftwerk.core.utils.XmlFileReader;
@@ -80,7 +81,23 @@ public class XMLReportingDataParser extends ReportingDataParser {
           String status = contactAttemptsElement.getFirstChildElement("status").getValue().toUpperCase();
           String timestamp = contactAttemptsElement.getFirstChildElement("date").getValue();
           reportingDataUE.addContactAttempts(new ContactAttempt(status, Long.parseLong(timestamp)));
-        }  
+        }
+
+      //Get identification
+      Element identificationElement = surveyUnitElement.getFirstChildElement("Identification");
+      reportingDataUE.setIdentification(new Identification());
+      if (identificationElement != null) {
+        if (identificationElement.getFirstChildElement(Constants.IDENTIFICATION_NAME) != null)
+          reportingDataUE.getIdentification().setIdentification(identificationElement.getFirstChildElement(Constants.IDENTIFICATION_NAME).getValue());
+        if (identificationElement.getFirstChildElement(Constants.ACCESS_NAME) != null)
+          reportingDataUE.getIdentification().setAccess(identificationElement.getFirstChildElement(Constants.ACCESS_NAME).getValue());
+        if (identificationElement.getFirstChildElement(Constants.SITUATION_NAME) != null)
+          reportingDataUE.getIdentification().setSituation(identificationElement.getFirstChildElement(Constants.SITUATION_NAME).getValue());
+        if (identificationElement.getFirstChildElement(Constants.CATEGORY_NAME) != null)
+          reportingDataUE.getIdentification().setCategory(identificationElement.getFirstChildElement(Constants.CATEGORY_NAME).getValue());
+        if (identificationElement.getFirstChildElement(Constants.OCCUPANT_NAME) != null)
+          reportingDataUE.getIdentification().setOccupant(identificationElement.getFirstChildElement(Constants.OCCUPANT_NAME).getValue());
+      }
       reportingData.addReportingDataUE(reportingDataUE);
     } 
     integrateReportingDataIntoUE(data, reportingData, withAllReportingData);